Ward against stack corruption in setMode
Something in how std.range.zip interacts with what we pass it gives us
semi-reproducible stack corruption on RusNet when we query channels for
modes. It's not reproducible outside of a live environment.

We can easily ward against it by detecting when the sign is no longer
'+' or '-' (generally it's 'ÿ').