Defensively strippedRight
353 RPL_NAMREPLY, 324 RPL_CHANNELMODEIS and MODE sometimes come with
trailing spaces, and others may too. 352 RPL_NAMREPLY definitely does,
but that is the text GECOS/real name field, so allow that.

Strip right on suspected types, and in parseGeneralCases on the defaulting
event.content set.

If it becomes a problem for a type, we'll just have to give it an entry
in parseSpecialcases, to exempt it from the strippedRight it would get
in parseGeneralCases.