v0.4.1
This is a minor patch release that merely adds an enum member
`IRCUser.Class.operator`.

It is instrumental to the upcoming versions of kameloso, which will
introduce the notion of a higher user class than whitelist, yet lower
than admin. As such this release is to be considered optional unless you
have a final switch over `IRCUser.Class`.

Other changes:

* Updated lu dependency to v0.2.1