Rework Postprocessor construction logic
So far we've kept an AliasSeq in postprocessors/package.d of all
Postprocessor classes, which we then foreached and instantiated.

This new approach, copied from an importing project, keeps an AliasSeq
of the string names of the modules containing said Postprocessors. These
are then imported and introspected for any classes that implement the
Postprocessor interface.