v0.9.0
This changes `IRCEvent.count` and `IRCEvent.altcount`, giving them
default values of `int.min` (a change from `int.init`). This allows us
to disambiguate an "unset" count vs a set count of 0. It obviously
collides at `int.min` now but that's inarguably a less common case than
`int.init` 0.

As this breaks expectations of the initial state of an `IRCEvent`, this
bumps the minor version.