v0.9.1
This is a small release adding a function `opEqualsCaseInsensitive`,
requested by importing projects. It allows for comparing two nicknames
case-insentively while still taking into account case mappings. As it is
merely an additive change, this only bumps the patch version.

Additionally all documentation was reworded to use DDOC REF macros to link
to other symbols, a change from backticks. This is to make it work better
with adrdox, which -- unlike ddox/scod -- doesn't parse text in backticks
as potential links.

No other changes.