v1.0.0
This release makes an incompatible change in the IRCEvent struct by
replacing the `count` and `altcount` ints with longs. As such this was
subject to a minor version bump, but since we've been sub-1.0 for so
long without any real changes, we might as well take the plunge.

Please file GitHub issues for any bugs you encounter.