v1.1.2
This is a maintenance release with only minor fixes and additions.

The one notable change is that a Twitch GLOBALUSERSTATE no longer sets
the user class to Class.admin. The original reasoning was that the client
should always be admin toward itself, but the logic doesn't hold on servers
(like Twitch) where several clients can be connected as the same
nickname, whom each have their sets of privileged users. An innocent
thing like a normal !echo command instantly lays the foundation for
serious abuse.

Importing projects can still emulate the old behaviour by listening for
GLOBALUSERSTATEs and assigning the class early themselves.