v1.4.0
This adds a new `IRCUser.Class.elevated` member due to requests from
importing projects. This is only a breaking change if you have a
`final switch` over `IRCUser.Class`, in which case you simply have to
add a new case.

This is otherwise mostly a patch with minor changes, but due to the
above we bump the minor version.