v1.5.0
This is an incremental update with only minor changes. However, API was
broken by changing the type of `IRCChannel.mods` from being a
`string[][char]` to being a `bool[string][char]`.

Originally, the variable housed string arrays of nicknames with a
particular mod as value in an associative array keyed by the mod prefix.
"prefix" in this context would be 'o' for `@`, 'v' for `+`, etc. This
meant that to find whether someone was included in the array, you had to
use `std.algorithm.searching.canFind`.

    auto ops = 'o' in someChannel.mods;
    if (!ops) return;

    if ((*ops).canFind(nickname))
    {
        // Nickname is an op.
    }

This was in contrast to the `bool[string]` `IRCChannel.users`, by which
you can determine whether or not a nickname is included by whether an
element with that nickname as key exists. The `bool` value is ignored.

Now `IRCChannel.mods` behave just like `IRCChannel.users`, and a simple
`opIn` can be used to tell if someone has a particular mod.

    auto ops = 'o' in someChannel.mods;
    if (!ops) return;

    if (nickname in ops)
    {
        // Nickname is an op.
    }

It has pros and cons, naturally, but ultimately it shouldn't impose that
big of a change for anyone. Feedback appreciated if it does.