v3.1.0
This release raises the `lu` dependency to a minimum of v3.2.0, and as
such bumps the minor version.

Added a `subchannel` member to `IRCEvent`. It's loosely defined, but the
idea is it refers to the secondary channel of an event, when there are
two. A concrete example would be Twitch's new shared channel chat, where
(when enabled) messages are repeated into one or more secondary channels.
In such a scenario, `IRCEvent.channel` would refer to the channel in
which the message effectively plays out, and `IRCEvent.subchannel` would
refer to the originating "source" channel.

Refactored the monolith that was `assertgen.d` into something more
modular and easier to maintain. It now also takes enough command-line
flags to be able to support reading tests from standard input, such as
when piping raw IRC lines to it. This was always possible when `--twitch`
was passed, as the prompts for server daemon, network and address were
then skipped, but this is possible for all networks now that these can
be specified at the command line.

Refactored `postprocessing/twitch.d` to be able to support the
`source-msg-id` tag. It now also is more careful when assigning values
to `IRCEvent` members, and will warn the user (more) if something is
overwritten.

* Reparsed all Twitch tests and updated the `tests/twitch.d` file with
  the modern parsing
* Added type `TWITCH_CHARITYDONATION`
* Balanced some Twitch tags so they don't overwrite each other
* Added a simple CI tool for running `dscanner` style checks on the codebase