v3.2.0
This release makes further changes to the structs in `dialect.defs`,
and as such bumps the minor version.

The `IRCEvent` string member `channel` is now an instance of struct
`IRCEvent.Channel`. In a normal build it only has a string member
`name`, which fills the same role as the previous `IRCEvent.channel`
string. In a Twitch build it additionally has a `ulong` member `id`,
which corresponds to the numeric user ID of the owner (broadcaster)
of a channel. This is a breaking change and will require use of
`IRCEvent.channel` to be string-replaced with `IRCEvent.channel.name`
in any existing code.

Compatibility methods were added to `IRCEvent.Channel` to help
transition to the new struct. Declare version
`ChannelMigrationCompatibility` to use them. It's not fully a drop-in
replacement to the old behaviour, but it should help with the transition.

Added an `altcontent` member to `IRCEvent`. This is a string that is
used for events where there are two separate message bodies; either
two chat messages, such as in a Twitch reply, or when the event is
accompanied by things like a system/server message. `IRCEvent.content`
makes up the primary body, and any secondary messages are to be stored
in `IRCEvent.altcontent`. This frees up `IRCEvent.aux[0]`, which was
becoming crowded for some events.

The Twitch-specific `IRCUser.id` is now a `ulong` instead of a `uint`.
This is potentially a breaking change, depending on how the value is
used in importing code. One solution is to migrate everything else to
`ulong` too; another is to cast the value to `uint` when needed.