v3.3.0
This fixes `library` and `bot` builds not being `pure` and `@safe`, which
we always assumed they were. It turns out `auto this() { /* ... */ }`
doesn't actually work. The fix is to mark constructors as `pure` and
`@safe` explicitly.

It's technically not a breaking change, but let's bump the minor
version to commemorate it; a fix to something that should never have been.